add entities and configurations for them

+++ Confectionery.API/Application/ViewModels/ConfectionViewModel.cs
@@ -5,7 +5,7 @@
         public string Name { get; set; }
         public string Description { get; set; }
         public decimal Price { get; set; }
-        public int Weight { get; set; }
+        public decimal Weight { get; set; }
         public bool IsOrderCountLimited { get; set; }
         public int MinimumOrderCount { get; set; }
         public bool isOutOfStock { get; set; }
